Interface changes:     Removed motorAxisPrimitive, motorAxisSetLogParam     motorAxisSetLog now takes a logParam parameter     drvMotorAsynConfigure takes an extra can_block parameter     Simulator create function only takes int parameters to avoid problems         passing double parameters from the vxWorks shell on PowerPC arch Functional changes:     Order of drvMotorAsyn interrupt callbacks has been changed to pass back         Float64 interrupts (typically position, etc.) before Int32 interrupts         (typically status), so that a move reaches its desired position before         it is signalled as complete. This is not a complete solution.     More parameter checking, particularly of axis number

+++ motorApp/MotorSimSrc/drvMotorSim.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif
 
-void motorSimCreate( int card, int axis, double hiLimit, double lowLimit, double home, int nCards, int nAxes );
+void motorSimCreate( int card, int axis, int hiLimit, int lowLimit, int home, int nCards, int nAxes );
 
 #ifdef __cplusplus
 }
